#17 [Refactor] MainView 에서 Calendar, WeatherDetail, PlanList 분리
